Factor out common logging code into a shared module
