+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
   "http-api",
   "org-node",
+  "shared",
 ]
 
 [patch.crates-io.radicle-keystore]
