+++ http-api/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Alexis Sellier <self@cloudhead.io>"]
 edition = "2018"
 
 [dependencies]
+shared = { path = "../shared", default-features = false }
 warp = { version = "0.3.1", features = ["tls"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["preserve_order"] }
@@ -21,8 +22,6 @@ either = { version = "1.6" }
 tracing = "0.1"
 tracing-subscriber = "0.2"
 async-trait = "0.1"
-tracing-stackdriver = { git = "https://github.com/adaszko/tracing-stackdriver.git", optional = true }
 
 [features]
-default = []
-gcp = ["tracing-stackdriver"]
+gcp = ["shared/gcp"]
