+++ org-node/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Alexis Sellier <self@cloudhead.io>"]
 edition = "2018"
 
 [dependencies]
+shared = { path = "../shared", default-features = false }
 async-trait = { version = "0.1" }
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1" }
@@ -23,8 +24,6 @@ tracing-subscriber = "0.2"
 ureq = { version = "2.1.1", features = ["json"] }
 ethers = { version = "0.4" }
 rustc-hex = { version = "2.1" }
-tracing-stackdriver = { git = "https://github.com/adaszko/tracing-stackdriver.git", optional = true }
 
 [features]
-default = []
-gcp = ["tracing-stackdriver"]
+gcp = ["shared/gcp"]
