+++ shared/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "shared"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+tracing = "0.1"
+tracing-subscriber = "0.2"
+tracing-stackdriver = { git = "https://github.com/adaszko/tracing-stackdriver.git", optional = true }
+
+[features]
+default = []
+gcp = ["tracing-stackdriver"]
