+++ shared/src/lib.rs
@@ -1,0 +1,3 @@
+mod logging;
+
+pub use logging::{init_logger, LogFmt};
